# server/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactCreate": {
        "properties": {
          "content": {
            "default": "",
            "maxLength": 1000000,
            "title": "Content",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "source": {
            "default": "human",
            "maxLength": 100,
            "minLength": 1,
            "title": "Source",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ArtifactType"
          }
        },
        "required": [
          "type",
          "title"
        ],
        "title": "ArtifactCreate",
        "type": "object"
      },
      "ArtifactResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "current_version": {
            "minimum": 1.0,
            "title": "Current Version",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ArtifactType"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "type",
          "title",
          "content",
          "current_version",
          "source",
          "created_at",
          "updated_at"
        ],
        "title": "ArtifactResponse",
        "type": "object"
      },
      "ArtifactRevisionCreate": {
        "properties": {
          "content": {
            "maxLength": 1000000,
            "title": "Content",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "source": {
            "default": "human",
            "maxLength": 100,
            "minLength": 1,
            "title": "Source",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "content"
        ],
        "title": "ArtifactRevisionCreate",
        "type": "object"
      },
      "ArtifactRevisionResponse": {
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ArtifactRevisionStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "artifact_id",
          "version",
          "title",
          "content",
          "source",
          "status",
          "created_at"
        ],
        "title": "ArtifactRevisionResponse",
        "type": "object"
      },
      "ArtifactRevisionStatus": {
        "enum": [
          "approved",
          "pending",
          "changes_requested",
          "rejected"
        ],
        "title": "ArtifactRevisionStatus",
        "type": "string"
      },
      "ArtifactType": {
        "enum": [
          "project_brief",
          "requirements",
          "acceptance_criteria",
          "research_note",
          "decision",
          "report",
          "code_diff",
          "generated_file",
          "external_link"
        ],
        "title": "ArtifactType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ClawChat Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/artifacts/{artifact_id}/revisions": {
      "post": {
        "operationId": "propose_artifact_revision_api_artifacts__artifact_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactRevisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Propose Artifact Revision",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/api/projects/{project_id}/artifacts": {
      "get": {
        "operationId": "list_project_artifacts_api_projects__project_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactResponse"
                  },
                  "title": "Response List Project Artifacts Api Projects  Project Id  Artifacts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Project Artifacts",
        "tags": [
          "artifacts"
        ]
      },
      "post": {
        "operationId": "create_project_artifact_api_projects__project_id__artifacts_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Project Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
...
  }
}